+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## [0.2.9] - Unreleased
+- Added iscc_compare function
 - Optimized soft_hash_audio performance
 - Removed Cython from build requirements
+- Fixed api listing tool
 - Updated dependencies
 
 ## [0.2.8] - 2022-04-21